remove tests covered in the learntorank library
